pyproject.toml
[project]
name = "twitter-notifications"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "orjson>=3.10.18",
    "pytelegrambotapi>=4.27.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "twscrape>=0.17.0",
]
